moving api types back to the virtual-assistant package
